[chore]Implement DRY on tests
