+++ testing/user_tests.py
@@ -1,0 +1,123 @@
+import unittest
+import json
+
+from webapi.routes import create_app, db
+
+class TestUserEndpoints(unittest.TestCase):
+    """Contains all user tests"""
+    def setUp(self):
+        """Set up the application with configurations for testing"""
+        self.app = create_app(config_name="testing")
+        self.tester = self.app.test_client(self)
+        with self.app.app_context():
+            db.create_all()
+            
+    def register_and_login(self, choice):
+        if choice == "both":
+            self.tester.post('/api/v2/auth/register',
+                                       data=dict(username = "admin",
+                                                 password = "1234",
+                                                 email = "test@email.com"))
+            tkn = self.tester.post('/api/v2/auth/login', data=dict(username = "admin",
+                                                                   password = "1234"))
+            self.token = json.loads(tkn.data.decode())['access-token']
+        
+        if choice == "login":
+            return self.tester.post('/api/v2/auth/login', data=dict(username = "admin",
+                                                                   password = "1234"))
+
+        if choice == "register":
+            return self.tester.post('/api/v2/auth/register',
+                                       data=dict(username = "admin",
+                                                 password = "1234",
+                                                 email = "test@email.com"))
+
+    def test_register(self):
+        """Test the register user endpoint"""
+        response = self.register_and_login("register")
+        self.assertEqual(response.status_code, 201)
+        self.assertIn("Registration successful", str(response.data))
+    
+    def test_register_noinput(self):
+        """Test a blank input on register endpoint"""
+        response = self.tester.post('/api/v2/auth/register',
+                               data=dict(username = "", password = "1234", email = "test@email.com"))
+        self.assertEqual(response.status_code, 400)
+        self.assertIn("Please insert", str(response.data))
+        
+    def test_already_registered(self):
+        """Test a user registering twice"""
+        self.register_and_login("register")
+        response = self.register_and_login("register")
+        self.assertEqual(response.status_code, 409)
+        self.assertIn("email already", str(response.data))
+
+    def test_login(self):
+        """Test the user login endpoint"""
+        self.register_and_login("register")
+        response = self.register_and_login("login")
+        self.assertEqual(response.status_code, 202)
+        self.assertIn("Logged in", str(response.data))
+    
+    def test_login_noinput(self):
+        """Test theres no input for login endpoint"""
+        self.register_and_login("register")
+        response = tester.post('/api/v2/auth/login',
+                               data=dict(username="", password="1234"))
+        self.assertEqual(response.status_code, 401)
+        self.assertIn("Could not verify", str(response.data))
+
+    def test_login_nouser(self):
+        """Test the user is not registered"""
+        response = self.register_and_login("login")
+        self.assertEqual(response.status_code, 401)
+        self.assertIn("Could not verify", str(response.data))
+        
+    def test_login_noinput(self):
+        """Test theres no input for login endpoint"""
+        self.register_and_login("register")
+        response = self.tester.post('/api/v2/auth/login',
+                               data=dict(username="admin", password="abcd"))
+        self.assertEqual(response.status_code, 401)
+        self.assertIn("Could not verify", str(response.data))
+
+    def test_logout(self):
+        """Test the logout user endpoint"""
+        self.register_and_login("both")
+        response = self.tester.post('/api/v2/auth/logout', headers={'x-access-token':self.token})
+        self.assertEqual(response.status_code, 202)
+        self.assertIn("logged out", str(response.data))
+        
+    def test_logout_twice(self):
+        """Test the user is already logged out"""
+        self.register_and_login("both")
+        self.tester.post('/api/v2/auth/logout', headers={'x-access-token':self.token})
+        response = self.tester.post('/api/v2/auth/logout', headers={'x-access-token':self.token})
+        self.assertEqual(response.status_code, 200)
+        self.assertIn("already logged out", str(response.data))
+
+    def test_reset_password(self):
+        """Test the reset password endpoint"""
+        self.register_and_login("both")
+        response = self.tester.post('/api/v2/auth/reset-password',
+                               headers={'x-access-token':self.token},
+                               data=dict(new_password = "somethingnew"))
+        self.assertEqual(response.status_code, 205)
+        self.assertIn("Password reset!", str(response.data))
+        
+    def test_reset_password_nologin(self):
+        """Test the reset password if user not logged in"""
+        response = self.register_and_login("both")
+        self.tester.post('/api/v2/auth/logout', headers={'x-access-token':self.token})
+        response = self.tester.post('/api/v2/auth/reset-password',
+                               headers={'x-access-token':self.token},
+                               data=dict(new_password = "somethingnew"))
+        self.assertEqual(response.status_code, 401)
+        self.assertIn("Please log in", str(response.data))
+
+    def tearDown(self):
+        """teardown all initialized variables."""
+        with self.app.app_context():
+            db.session.remove()
+            db.drop_all()
+
